@@ -27,42 +27,40 @@
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.40.5",
-    "@octokit/core": "^5.1.0",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "@rollup/plugin-typescript": "^11.1.6",
-    "@types/chai": "^4.3.11",
-    "@types/mocha": "^10.0.6",
-    "@types/node": "^20.11.19",
-    "@types/sinon": "^17.0.3",
+    "@microsoft/api-extractor": "^7.52.5",
+    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@rollup/plugin-typescript": "^12.1.2",
+    "@types/chai": "^4.3.17",
+    "@types/mocha": "^10.0.7",
+    "@types/node": "^22.15.3",
+    "@types/sinon": "^17.0.4",
     "@types/sinon-chai": "^3.2.12",
-    "@typescript-eslint/eslint-plugin": "^7.0.2",
-    "@typescript-eslint/parser": "^7.0.2",
+    "@typescript-eslint/eslint-plugin": "^8.31.1",
+    "@typescript-eslint/parser": "^8.31.1",
     "benchmark": "^2.1.4",
-    "chai": "^4.3.10",
+    "chai": "^4.4.1",
     "chalk": "^5.3.0",
-    "dbx-js-tools": "github:mongodb-js/dbx-js-tools",
-    "eslint": "^8.56.0",
-    "eslint-config-prettier": "^9.1.0",
+    "dbx-js-tools": "github:mongodb-js/dbx-js-tools#main",
+    "eslint": "^9.25.1",
+    "eslint-config-prettier": "^10.1.2",
     "eslint-plugin-no-bigint-usage": "file:etc/eslint/no-bigint-usage",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-tsdoc": "^0.2.17",
-    "magic-string": "^0.30.7",
-    "mocha": "10.3.0",
+    "eslint-plugin-prettier": "^5.2.6",
+    "eslint-plugin-tsdoc": "^0.4.0",
+    "magic-string": "^0.30.11",
+    "mocha": "^10.7.0",
     "node-fetch": "^3.3.2",
     "nyc": "^15.1.0",
-    "prettier": "^3.2.5",
-    "rollup": "^4.12.0",
-    "sinon": "^17.0.1",
+    "prettier": "^3.5.3",
+    "rollup": "^4.40.1",
+    "sinon": "^18.0.0",
     "sinon-chai": "^3.7.0",
     "source-map-support": "^0.5.21",
-    "standard-version": "^9.5.0",
-    "tar": "^6.2.0",
+    "tar": "^7.4.3",
     "ts-node": "^10.9.2",
-    "tsd": "^0.30.5",
-    "typescript": "^5.0.4",
+    "tsd": "^0.31.1",
+    "typescript": "^5.8.3",
     "typescript-cached-transpile": "0.0.6",
-    "uuid": "^9.0.1"
+    "uuid": "^11.1.0"
   },
   "tsd": {
     "directory": "test/types",
