@@ -62,53 +62,53 @@
     }
   },
   "devDependencies": {
+    "@aws-sdk/credential-providers": "^3.632.0",
     "@iarna/toml": "^2.2.5",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.36.4",
-    "@microsoft/tsdoc-config": "^0.16.2",
-    "@mongodb-js/zstd": "^1.1.0",
-    "@octokit/core": "^4.2.4",
-    "@types/chai": "^4.3.5",
-    "@types/chai-subset": "^1.3.3",
-    "@types/express": "^4.17.17",
-    "@types/kerberos": "^1.1.2",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^20.5.9",
-    "@types/saslprep": "^1.0.1",
-    "@types/semver": "^7.5.0",
-    "@types/sinon": "^10.0.16",
-    "@types/sinon-chai": "^3.2.9",
-    "@types/whatwg-url": "^11.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
-    "chai": "^4.3.7",
+    "@microsoft/api-extractor": "^7.52.5",
+    "@microsoft/tsdoc-config": "^0.17.1",
+    "@mongodb-js/zstd": "^2.0.1",
+    "@types/chai": "^4.3.17",
+    "@types/chai-subset": "^1.3.5",
+    "@types/express": "^5.0.1",
+    "@types/kerberos": "^1.1.5",
+    "@types/mocha": "^10.0.9",
+    "@types/node": "^22.15.3",
+    "@types/saslprep": "^1.0.3",
+    "@types/semver": "^7.7.0",
+    "@types/sinon": "^17.0.4",
+    "@types/sinon-chai": "^4.0.0",
+    "@types/whatwg-url": "^13.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.31.1",
+    "@typescript-eslint/parser": "^8.31.1",
+    "chai": "^4.4.1",
     "chai-subset": "^1.6.0",
     "chalk": "^4.1.2",
-    "eslint": "^8.48.0",
-    "eslint-config-prettier": "^8.10.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
-    "eslint-plugin-tsdoc": "^0.2.17",
-    "eslint-plugin-unused-imports": "^2.0.0",
-    "express": "^4.18.2",
-    "gcp-metadata": "^5.2.0",
+    "eslint": "^9.25.1",
+    "eslint-config-prettier": "^10.1.2",
+    "eslint-plugin-mocha": "^10.4.1",
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-simple-import-sort": "^12.1.1",
+    "eslint-plugin-tsdoc": "^0.4.0",
+    "eslint-plugin-unused-imports": "^4.1.4",
+    "express": "^5.1.0",
+    "gcp-metadata": "^5.3.0",
     "js-yaml": "^4.1.0",
-    "mocha": "^10.2.0",
+    "mocha": "^10.8.2",
     "mocha-sinon": "^2.1.2",
-    "mongodb-client-encryption": "^6.0.0",
-    "mongodb-legacy": "^6.0.0",
+    "mongodb-client-encryption": "^6.4.0",
+    "mongodb-legacy": "^6.1.3",
     "nyc": "^15.1.0",
-    "prettier": "^2.8.8",
-    "semver": "^7.5.4",
-    "sinon": "^15.2.0",
+    "prettier": "^3.5.3",
+    "semver": "^7.7.0",
+    "sinon": "^18.0.1",
     "sinon-chai": "^3.7.0",
     "snappy": "^7.2.2",
-    "socks": "^2.7.1",
+    "socks": "^2.8.1",
     "source-map-support": "^0.5.21",
-    "ts-node": "^10.9.1",
-    "tsd": "^0.28.1",
-    "typescript": "^5.0.4",
+    "ts-node": "^10.9.2",
+    "tsd": "^0.32.0",
+    "typescript": "5.8.3",
     "typescript-cached-transpile": "^0.0.6",
     "v8-heapsnapshot": "^1.3.1",
     "yargs": "^17.7.2"
