@@ -17,58 +17,59 @@
     "orm",
     "db"
   ],
+  "type": "commonjs",
   "license": "MIT",
   "dependencies": {
-    "bson": "^6.2.0",
-    "kareem": "2.5.1",
-    "mongodb": "6.3.0",
+    "bson": "^6.10.4",
+    "kareem": "2.6.3",
+    "mongodb": "~6.18.0",
     "mpath": "0.9.0",
     "mquery": "5.0.0",
     "ms": "2.1.3",
-    "sift": "16.0.1"
+    "sift": "17.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.24.0",
-    "@babel/preset-env": "7.24.0",
-    "@typescript-eslint/eslint-plugin": "^6.2.1",
-    "@typescript-eslint/parser": "^6.2.1",
-    "acquit": "1.3.0",
+    "@babel/core": "7.28.0",
+    "@babel/preset-env": "7.28.0",
+    "@mongodb-js/mongodb-downloader": "^0.4.2",
+    "@typescript-eslint/eslint-plugin": "^8.19.1",
+    "@typescript-eslint/parser": "^8.19.1",
+    "acquit": "1.4.0",
     "acquit-ignore": "0.2.1",
     "acquit-require": "0.1.1",
+    "ajv": "8.17.1",
     "assert-browserify": "2.0.0",
-    "axios": "1.1.3",
     "babel-loader": "8.2.5",
-    "benchmark": "2.1.4",
     "broken-link-checker": "^0.7.8",
     "buffer": "^5.6.0",
-    "cheerio": "1.0.0-rc.12",
-    "crypto-browserify": "3.12.0",
-    "dotenv": "16.4.5",
+    "cheerio": "1.1.2",
+    "crypto-browserify": "3.12.1",
     "dox": "1.0.0",
-    "eslint": "8.57.0",
-    "eslint-plugin-markdown": "^3.0.1",
-    "eslint-plugin-mocha-no-only": "1.1.1",
-    "express": "^4.18.1",
-    "fs-extra": "~11.2.0",
-    "highlight.js": "11.8.0",
+    "eslint": "8.57.1",
+    "eslint-plugin-markdown": "^5.1.0",
+    "eslint-plugin-mocha-no-only": "1.2.0",
+    "express": "^4.19.2",
+    "fs-extra": "~11.3.0",
+    "highlight.js": "11.11.1",
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
-    "markdownlint-cli2": "^0.12.1",
-    "marked": "4.3.0",
+    "markdownlint-cli2": "^0.18.1",
+    "marked": "15.0.12",
     "mkdirp": "^3.0.1",
-    "mocha": "10.3.0",
-    "moment": "2.x",
-    "mongodb-memory-server": "8.15.1",
+    "mocha": "11.7.1",
+    "moment": "2.30.1",
+    "mongodb-memory-server": "10.1.4",
+    "mongodb-runner": "^5.8.2",
     "ncp": "^2.0.0",
     "nyc": "15.1.0",
-    "pug": "3.0.2",
+    "pug": "3.0.3",
     "q": "1.5.1",
-    "sinon": "17.0.1",
+    "sinon": "21.0.0",
     "stream-browserify": "3.0.0",
-    "tsd": "0.30.7",
-    "typescript": "5.3.3",
-    "uuid": "9.0.1",
-    "webpack": "5.90.3"
+    "tsd": "0.32.0",
+    "typescript": "5.8.3",
+    "uuid": "11.1.0",
+    "webpack": "5.101.0"
   },
   "directories": {
     "lib": "./lib/mongoose"
