@@ -1,7 +1,7 @@
 {
   "name": "sift",
   "description": "MongoDB query filtering in JavaScript",
-  "version": "16.0.1",
+  "version": "17.1.3",
   "repository": "crcn/sift.js",
   "sideEffects": false,
   "author": {
@@ -20,19 +20,18 @@
     "@rollup/plugin-replace": "^2.3.2",
     "@rollup/plugin-typescript": "8.2.1",
     "@types/node": "^13.7.0",
-    "bson": "^4.0.3",
-    "eval": "^0.1.4",
-    "husky": "^1.2.1",
-    "immutable": "^3.7.6",
-    "mocha": "8.3.2",
+    "bson": "^6.6.0",
+    "eval": "^0.1.8",
+    "husky": "^9.0.11",
+    "mocha": "10.4.0",
     "mongodb": "^3.6.6",
-    "prettier": "1.15.3",
-    "pretty-quick": "^1.11.1",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.7.2",
-    "rollup-plugin-terser": "^7.0.2",
-    "tslib": "2.2.0",
-    "typescript": "4.2.4"
+    "prettier": "3.2.5",
+    "pretty-quick": "^4.0.0",
+    "rimraf": "^5.0.5",
+    "rollup": "^4.14.2",
+    "@rollup/plugin-terser": "^0.4.4",
+    "tslib": "2.6.2",
+    "typescript": "5.4.5"
   },
   "main": "./index.js",
   "module": "./es5m/index.js",
