@@ -1,35 +1,38 @@
-// const express = require('express')// method-1
-import express from "express"; // method-2
+import express from "express";
 import dotenv from "dotenv"; 
 import connectDB from "./config/database.js";
 import userRoute from "./routes/userRoute.js";
 import messageRoute from "./routes/messageRoute.js";
 import cookieParser from "cookie-parser";
 import cors from "cors";
-import { app,server } from "./socket/socket.js";
-dotenv.config({});
+import { app, server } from "./socket/socket.js";
 
- 
-const PORT = process.env.PORT || 5000;
+dotenv.config();
+
+const PORT = process.env.PORT || 8080;
 
 // middleware
-app.use(express.urlencoded({extended:true}));
+app.use(express.urlencoded({ extended: true }));
 app.use(express.json()); 
 app.use(cookieParser());
-const corsOption={
-    origin:'http://localhost:3000',
-    credentials:true
+
+const corsOption = {
+    origin: "http://localhost:3000",
+    credentials: true,
 };
 app.use(cors(corsOption)); 
 
-
-// routes
-app.use("/api/v1/user",userRoute); 
-app.use("/api/v1/message",messageRoute);
- 
-
-server.listen(PORT, ()=>{
-    connectDB();
-    console.log(`Server listen at prot ${PORT}`);
+// test route
+app.get("/", (req, res) => {
+    res.send("Backend is running fine 🚀");
 });
 
+// routes
+app.use("/api/v1/user", userRoute); 
+app.use("/api/v1/message", messageRoute);
+
+// start server
+server.listen(PORT, () => {
+    connectDB();
+    console.log(`Server running on http://localhost:${PORT}`);
+});
