@@ -8,20 +8,16 @@ const Messages = () => {
     useGetMessages();
     useGetRealTimeMessage();
     const { messages } = useSelector(store => store.message);
+
+    console.log("messages value =>", messages); // for debugging
+
     return (
         <div className='px-4 flex-1 overflow-auto'>
-            {
-               messages && messages?.map((message) => {
-                    return (
-                        <Message key={message._id} message={message} />
-                    )
-                })
-            }
-
+            {Array.isArray(messages) && messages.map((message) => (
+                <Message key={message._id} message={message} />
+            ))}
         </div>
-
-
     )
 }
 
-export default Messages
+export default Messages;
